Move DocComment into DocComment namespace
This is in preparation for adding the following:

* DocCommentInterface
* DocCommentFactory
* DocCommentFactoryInterface

Yes...it complicates a nice, clean, simple system. But it also makes things
testable!
